refactor(ThemeContext): add Providers component
