+++ src/context/ThemeContext.ts
@@ -1,0 +1,14 @@
+import {createContext} from "react";
+import THEMES from "@/THEMES.json";
+import {TThemeName} from "@/types/TThemeName";
+
+const initialChangeThemeToFn = (newThemeName: TThemeName) => {}
+
+const initialContextValue = {
+    currentTheme: THEMES['default'],
+changeThemeTo: initialChangeThemeToFn
+}
+
+const ThemeContext = createContext(initialContextValue);
+
+export default ThemeContext;
